+++ packages/cli/__tests__/zosfiles/__integration__/compare/ds/cli.files.compare.ds.integration.test.ts
@@ -45,5 +45,4 @@ describe("Compare two data sets", () => {
         expect(response.stderr.toString()).toContain("Missing Positional Argument");
         expect(response.stderr.toString()).toContain("dataSetName1");
     });
-
 });