+++ packages/cli/src/zosfiles/-strings-/en.ts
@@ -318,27 +318,23 @@ export default {
                 DESCRIPTION: "Download all data sets that match a DSLEVEL pattern (see help below). " +
                     "You can use several options to qualify which data sets will be skipped and how the " +
                     "downloaded files will be structured. Data sets that are neither physical sequential nor " +
-                    "partitioned data sets (with members) will be excluded. ",
+                    "partitioned data sets (with members) will be excluded.",
                 POSITIONALS: {
-                    PATTERN: `The pattern or patterns to match data sets against. Also known as 'DSLEVEL'. The
-                    following special sequences can be used in the pattern:
+                    PATTERN: `The pattern or patterns to match data sets against. Also known as 'DSLEVEL'. The following special sequences can be ` +
+                    `used in the pattern:
                     ${TextUtils.chalk.yellow("%")}: matches any single character
-                    ${TextUtils.chalk.yellow("*")}: matches any number of characters within a data set name qualifier (e.g.
-                    "ibmuser.j*.old" matches "ibmuser.jcl.old" but not "ibmuser.jcl.very.old")
-                    ${TextUtils.chalk.yellow("**")}: matches any number of characters within any number of data set name
-                    qualifiers (e.g. "ibmuser.**.old" matches both "ibmuser.jcl.old" and
-                    "ibmuser.jcl.very.old")
-                    However, the pattern cannot begin with any of these sequences.You can specify
-                    multiple patterns separated by commas, for example
-                    "ibmuser.**.cntl,ibmuser.**.jcl"`
+                    ${TextUtils.chalk.yellow("*")}: matches any number of characters within a data set name qualifier ` +
+                    `(e.g. "ibmuser.j*.old" matches "ibmuser.jcl.old" but not "ibmuser.jcl.very.old")
+                    ${TextUtils.chalk.yellow("**")}: matches any number of characters within any number of data set name qualifiers ` +
+                    `(e.g. "ibmuser.**.old" matches both "ibmuser.jcl.old" and "ibmuser.jcl.very.old")
+                    However, the pattern cannot begin with any of these sequences.You can specify multiple patterns separated by commas, ` +
+                    `for example "ibmuser.**.cntl,ibmuser.**.jcl"`
                 },
                 EXAMPLES: {
-                    EX1: `Download all data sets beginning with "ibmuser" and ending with
-                    ".cntl" or ".jcl" to the local directory "jcl" to files with the extension ".jcl"`,
-                    EX2: `Download all data sets that begin with "ibmuser.public.project" or
-                    "ibmuser.project.private", excluding those that end in "lib" to the local
-                    directory "project", providing a custom mapping of data set low level qualifier
-                    to local file extension`
+                    EX1: `Download all data sets beginning with "ibmuser" and ending with ".cntl" or ".jcl" to the local directory "jcl" to ` +
+                        `files with the extension ".jcl"`,
+                    EX2: `Download all data sets that begin with "ibmuser.public.project" or "ibmuser.project.private", excluding those that end ` +
+                    `in "lib" to the local directory "project", providing a custom mapping of data set low level qualifier to local file extension`
                 }
             }
         },
@@ -719,8 +715,9 @@ export default {
             VOLUME: "The volume serial (VOLSER) where the first data set resides. You can use this option at any time. However, the " +
                 "VOLSER is required only when the data set is not cataloged on the system. A VOLSER is analogous to a drive name on a PC.",
             VOLUME2: "The volume serial (VOLSER) where the second data set resides.",
-            NOSEQNUM: "If you are comparing two files that contain sequence numbers, this option strips the last eight " +
-                "characters from each string.",
+            SEQNUM: "If you are comparing two files that contain sequence numbers, this option controls if the sequences numbers are " +
+                "removed from the end of each string. The default is to keep the sequence numbers. Use the --no-seqnum option to not " +
+                "include them.",
             CONTEXTLINES: "The number of context lines that display before and after detected non-matching lines. By default all matching " +
                 "lines display. If you want to limit the amount of data returned to only lines with differences use the context " +
                 "lines option to reduce the matching lines to only those before and after non-matching lines. Using the value " +