+++ packages/cli/src/zosfiles/compare/ds/Dataset.handler.ts
@@ -72,7 +72,7 @@ export default class DatasetHandler extends ZosFilesBaseHandler {
         let dsContentString1 = "";
         let dsContentString2 = "";
 
-        if (commandParameters.arguments.noseqnum) {
+        if(commandParameters.arguments.seqnum == false){
             const seqnumlen = 8;
 
             const dsStringArray1 = dsContentBuf1.toString().split("\n");